chore: refactor, cleanup
* compose middleware
* named exports for middleware
* clean up control flow
* lift constants
* strict mode
* enforce type checking, optional chaining

+++ src/controllers/todo.ts
@@ -1,4 +1,4 @@
-import { PrismaClient, Todo, User, UserWhereInput } from '@prisma/client'
+import { PrismaClient, Todo, UserWhereInput } from '@prisma/client'
 import { Request, Response } from 'express'
 
 const prisma = new PrismaClient()
@@ -10,21 +10,22 @@ export const TodoController = {
   //!    category: String (required): work, personal
   //!    content: String (required): take out trash
   //! =================================================
-  create: async (req: Request, res: Response) => {
+  createOne: async (req: Request, res: Response) => {
     const { id } = res.locals.payload
+
     const { category, content } = req.body
+    const data = { content, category, author: { connect: { id } } }
 
     if (!(content && category)) {
-      res.status(400).json({ err: 'Must provide content and category' })
+      return res.status(400).json({ err: 'Must provide content and category' })
     }
 
-    let todo: Todo
+    let todo: Todo | null = null
 
     try {
-      const data = { content, category, author: { connect: { id } } }
       todo = await prisma.todo.create({ data })
     } catch (err) {
-      res.status(401).json({ err: 'Unable to create new todo' })
+      return res.status(401).json({ err: 'Unable to create new todo' })
     }
 
     res.status(200).json({ todo })
@@ -36,17 +37,20 @@ export const TodoController = {
   findAll: async (_req: Request, res: Response) => {
     const { id } = res.locals.payload
 
-    let user: UserWhereInput
+    const where = { id: Number(id) }
+    const include = { todos: true }
+
+    let user: UserWhereInput | null = null
 
     try {
-      const where = { id: Number(id) }
-      const include = { todos: true }
       user = await prisma.user.findOne({ where, include })
     } catch (err) {
-      res.status(401).send()
+      return res
+        .status(401)
+        .json({ err: 'You are not the author of this todo' })
     }
 
-    const { todos } = user
+    const todos = user?.todos
 
     res.status(200).json({ todos })
   },
@@ -56,20 +60,24 @@ export const TodoController = {
   //!
   //! If no query string is provided, return all todos
   //! =================================================
-  filterBy: async (req: Request, res: Response) => {
+  filterMany: async (req: Request, res: Response) => {
     const { id } = res.locals.payload
 
     const { search }: { search?: string } = req.query
+    const contains = { contains: search }
+    const OR = [{ category: contains }, { content: contains }]
+    const author = { id: Number(id) }
 
-    let todos: Todo[]
+    let todos: Todo[] | null = null
 
     try {
-      const contains = { contains: search }
-      const OR = [{ category: contains }, { content: contains }]
-      const author = { id: Number(id) }
       todos = await prisma.todo.findMany({ where: { author, OR } })
     } catch (err) {
-      res.status(401).json({ err: 'No todos by that author' })
+      return res.status(500).json({ err: 'Error finding todos' })
+    }
+
+    if (!todos) {
+      return res.status(400).json({ err: 'No todos by that author' })
     }
 
     res.status(200).json({ todos })
@@ -78,18 +86,69 @@ export const TodoController = {
   //! =================================================
   //! Delete a todo by id
   //! =================================================
-  deleteBy: async (req: Request, res: Response) => {
+  deleteOne: async (req: Request, res: Response) => {
     const { id } = req.params
 
-    let todo: Todo
+    const userId = res.locals.payload.id
+
+    const where = { id: Number(id) }
+
+    let todo: Todo | null = null
+    let owner: Todo | null = null
 
     try {
-      const where = { id: Number(id) }
+      owner = await prisma.todo.findOne({ where })
+    } catch (err) {
+      return res.status(500).json({ err: 'Problem finding todo' })
+    }
+
+    if (!(owner?.authorId === userId)) {
+      return res
+        .status(401)
+        .json({ err: 'You are not the author of this todo' })
+    }
+
+    try {
       todo = await prisma.todo.delete({ where })
     } catch (err) {
-      res.status(401).json({ err: 'No todo with that ID' })
+      return res.status(500).json({ err: 'Problem deleting todo' })
+    }
+
+    if (!todo) {
+      return res.status(401).json({ err: 'No todo with that ID' })
     }
 
     res.status(202).json({ todo })
+  },
+
+  updateOne: async (req: Request, res: Response) => {
+    const { id } = req.params
+
+    const { category, content } = req.body
+    const where = { id: Number(id) }
+
+    if (!(category && content)) {
+      return res.status(400).json({ err: 'Must provide category and content' })
+    }
+
+    let todo: Todo | null = null
+
+    try {
+      todo = await prisma.todo.findOne({ where })
+    } catch (err) {
+      return res.status(500).json({ err: 'Failed to update todo' })
+    }
+
+    if (!todo) {
+      return res.status(400).json({ err: 'Todo with that id does not exist' })
+    }
+
+    try {
+      await prisma.todo.update({ where, data: { category, content } })
+    } catch (err) {
+      return res.status(500).json({ err: 'Failed to update todo' })
+    }
+
+    res.status(204).send()
   },
 }
